+++ landscape-architecture.WebAPI/landscape-architecture.WebAPI.tests/Controllers/UsersControllerUnitTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace landscape_architecture.WebAPI.Tests.Controllers
+{
+    internal class UsersControllerUnitTests
+    {
+    }
+}
