+++ landscape-architecture.WebAPI/landscape-architecture.WebAPI.tests/landscape-architecture.WebAPI.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
